[O2B-1057] extracted copyLink into a seperate component
