+++ lib/public/components/common/selection/infoLoggerButtonGroup/infoLoggerButtonGroup.js
@@ -11,32 +11,12 @@
  *  or submit itself to any jurisdiction.
  */
 
-import { frontLink } from '../navigation/frontLink.js';
-import { dropdown } from '../popover/dropdown.js';
+import { frontLink } from '../../navigation/frontLink.js';
+import { dropdown } from '../../popover/dropdown.js';
 import { iconCaretBottom } from '/js/src/icons.js';
 import { h } from '/js/src/index.js';
-import { getInfologgerLink } from '../../../services/externalRouting/getInfologgerLink.js';
-import { clipboardTarget } from '/components/common/clipboardTarget.js';
-
-/**
- * Copies a detail page link to the clipboard based on the given parameters
- *
- * @param {string} key the key that specifies what detail page to link to
- * @param {string} value the value that specifies what detail page to link to
- * @param {string} page the page that specifies what detail page to link to
- * @return {clipboardTarget} the copy link anchor tag wrapped in a clipboardTarget component
- */
-const copyLink = (key, value, page) => {
-    const url = `${window.location.origin}/?page=${page}&${key}=${value}`;
-
-    const component = h('a', {
-        id: `copy-${value}`,
-        style: 'cursor: pointer',
-    }, 'Copy Link');
-
-    return clipboardTarget(url, component, () => { });
-};
-
+import { getInfologgerLink } from '../../../../services/externalRouting/getInfologgerLink.js';
+import { copyDetailPageLink } from './copyDetailPageLink.js';
 /**
  * Renders a button group that links to a specific details page based on the given parameters, 
  * and a dropdown menu with options to copy the link or access the related infologgers.
@@ -60,7 +40,7 @@ export const infoLoggerButtonGroup = (parameters) => {
                 h(
                     '.flex-column.p2.g3',
                     [
-                        copyLink(Object.keys(key)[0], value, page),
+                        copyDetailPageLink(Object.keys(key)[0], value, page),
                         flpInfologger && h('a', { href: flpInfologger, target: '_blank' }, 'Infologger FLP'),
                         epnInfologger && h('a', { href: epnInfologger, target: '_blank' }, 'Infologger EPN'),
                     ],
