+++ lib/public/views/Environments/ActiveColumns/environmentsActiveColumns.js
@@ -14,7 +14,7 @@
 import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.js';
 import { formatRunsList } from '../../Runs/format/formatRunsList.js';
 import { displayEnvironmentStatus } from '../format/displayEnvironmentStatus.js';
-import { infoLoggerButtonGroup } from '../../../components/common/selection/infoLoggerButtonGroup.js';
+import { infoLoggerButtonGroup } from '../../../components/common/selection/infoLoggerButtonGroup/infoLoggerButtonGroup.js';
 
 /**
  * List of active columns for a generic Environments component
